Changes from code review
